강민정 : 인벤토리 카드 1차 구현

+++ NoUmbrellasAllowed/CardTotalPrice.cpp
@@ -74,6 +74,6 @@ int ACardTotalPrice::TotalPriceCheck()
         }
 
     }
-    TotalPrice = (TotalPrice + TexturePlusPrice) * ConditionPercent;
+    TotalPrice = static_cast<int>((TotalPrice + TexturePlusPrice) * ConditionPercent);
     return TotalPrice;
 }
